Create RestoreSummaryRequest objects in parallel
When restoring a lot of projects, the creation of RestoreSummaryRequest objects is slow.  By creating them in parallel large restores are 15% faster.

I also had to update the return value of IPreLoadedRestoreRequestProvider.CreateRequests to IReadOnlyCollection since I'm using a ConcurrentBag<T> in the Parallel.ForEach which doesn't support IList<T>